# Faster LOS calc signal with new_calc = True