@lasofivec 2020-01-07
Not sure of what you mean by "re assigning" but, the problem came from the fact that the result stored in `sig` was C-contiguous while `sig` is explicitly defined as a Fortran contiguous array (faster computation for all other cases).
So converting to result to a F-contiguous array solves the problem: `sig = np.asfortranarray(...`