@lasofivec 2020-01-07
Latest results:

```
In [1]: import tofu as tf
/home/ITER/mendozl/repositories/tofu/usr/lib/python3.6/site-packages/tofu-1.4.2a5-py3.6-linux-x86_64.egg/tofu/imas2tofu/__init__.py:83: UserWarning:
You do not seem to be using the latest IMAS version:
'module list' vs 'module av IMAS' suggests:
	- Current version: 3.24.0-4.1.5
	- Latest version : 3.25.0-4.4.0
  warnings.warn(msg)
/home/ITER/mendozl/repositories/tofu/usr/lib/python3.6/site-packages/tofu-1.4.2a5-py3.6-linux-x86_64.egg/tofu/__init__.py:95: UserWarning:
The following subpackages are not available:
    - tofu.mag
  => see tofu.dsub[<subpackage>] for details.
  warnings.warn(msg)

In [2]: multi = tf.imas2tofu.MultiIDSLoader(user='hoeneno', tokamak='convert', shot=134000, run=29, ids=['edge_sources'])
Getting ids   [occ]  tokamak  user     version  shot    run  refshot  refrun
------------  -----  -------  -------  -------  ------  ---  -------  ------
edge_sources  [0]    convert  hoeneno  3        134000  29   -1       -1

In [3]: _dshort = {'core_sources': {'1drhotn':'source[identifier.name=radiation].profiles_1d[time].grid.rho_tor_norm',
   ...:                             '1deEnergy':'source[identifier.name=radiation].profiles_1d[time].electrons.energy'
   ...:                                  },
   ...:            'equilibrium': {'2dpsi': 'time_slice[time].profiles_2d[0].psi',
   ...:                            '2dmeshR': 'time_slice[time].profiles_2d[0].r',
   ...:                            '2dmeshZ': 'time_slice[time].profiles_2d[0].z'}}
   ...:

In [4]: multi.set_shortcuts(dshort=_dshort)

In [5]: plasma = multi.to_Plasma2D()

In [6]: cam = tf.load('/home/ITER/munechk/public/MyTofu/output/ITER_test_camera_config.npz')
Loaded from:
    /home/ITER/munechk/public/MyTofu/output/ITER_test_camera_config.npz

In [7]: %timeit  sig_oldufunc, units = cam.calc_signal_from_Plasma2D(plasma, quant='edge_sources.2dradiation', plot=False, res=0.1, newcalc=False, fill_value=0.)
7.44 s ± 114 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [8]: %timeit sig_newbasic, units = cam.calc_signal_from_Plasma2D(plasma, quant='edge_sources.2dradiation', plot=False, res=0.1, newcalc=True, method='sum', minimize='calls')
6.73 s ± 102 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```