@lasofivec 2020-01-07
If tests pass and you agree. I am ok for merging this PR