@lasofivec 2020-01-07
Yes :) it looks like everything is working !